refactor(node): create _test_utils.ts
