+++ node/_fs/_fs_chmod_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 import { assert, fail } from "../../testing/asserts.ts";
 import { isWindows } from "../../_util/os.ts";
-import { assertCallbackErrorUncaught } from "../_utils.ts";
+import { assertCallbackErrorUncaught } from "../_test_utils.ts";
 import { chmod, chmodSync } from "./_fs_chmod.ts";
 
 Deno.test({
